requirements.txt
openai==1.30.1
python-dotenv==1.0.1
pypdf==4.2.0
pyperclip==1.8.2 